+++ app/src/main/java/com/depromeet/threedays/front/domain/validation/TokenValidator.java
@@ -1,23 +1,15 @@
-package com.depromeet.threedays.front.support;
+package com.depromeet.threedays.front.domain.validation;
 
 import com.depromeet.threedays.front.client.AuthClient;
 import com.depromeet.threedays.front.client.model.KeyProperties;
-import com.depromeet.threedays.front.client.model.KeyProperties.KeyProperty;
 import com.depromeet.threedays.front.client.property.auth.AppleAuthProperty;
 import com.depromeet.threedays.front.config.security.filter.token.IdTokenProperties;
 import com.depromeet.threedays.front.config.security.filter.token.TokenResolver;
 import com.depromeet.threedays.front.exception.ExternalIntegrationException;
 import com.depromeet.threedays.front.exception.JsonParsingException;
 import com.depromeet.threedays.front.web.request.member.AppleSignMemberRequest;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.nimbusds.jose.JWSVerifier;
-import com.nimbusds.jose.crypto.RSASSAVerifier;
-import com.nimbusds.jose.jwk.JWK;
-import com.nimbusds.jose.jwk.RSAKey;
-import com.nimbusds.jwt.SignedJWT;
 import java.net.URI;
 import java.net.URISyntaxException;
-import java.security.interfaces.RSAPublicKey;
 import java.util.Date;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -37,7 +29,7 @@ public class TokenValidator {
 		KeyProperties keyProperties = getKeyProperties(property);
 
 		IdTokenProperties idTokenProperties =
-			tokenResolver.extractPropertiesByToken(request.getIdToken());
+				tokenResolver.extractPropertiesByToken(request.getIdToken());
 
 		Date currentTime = new Date(System.currentTimeMillis());
 		if (!currentTime.before(idTokenProperties.getExp())) {
@@ -55,11 +47,6 @@ public class TokenValidator {
 		if (!property.getClientId().equals(idTokenProperties.getAud())) {
 			throw new JsonParsingException("token.not.valid");
 		}
-
-		// todo publicKey 검증 방법 조사
-		if (!verifyPublicKey(keyProperties, request.getIdToken())) {
-			throw new JsonParsingException("token.not.valid");
-		}
 	}
 
 	private KeyProperties getKeyProperties(AppleAuthProperty property) {
@@ -70,25 +57,9 @@ public class TokenValidator {
 		}
 	}
 
-	private boolean verifyPublicKey(KeyProperties keyProperties, String idToken) {
-
-		try {
-			SignedJWT signedJWT = SignedJWT.parse(idToken);
-
-			for (KeyProperty key : keyProperties.getKeys()) {
-				ObjectMapper objectMapper = new ObjectMapper();
-				RSAKey rsaKey = (RSAKey) JWK.parse(objectMapper.writeValueAsString(key));
-				RSAPublicKey publicKey = rsaKey.toRSAPublicKey();
-				JWSVerifier verifier = new RSASSAVerifier(publicKey);
-
-				if (signedJWT.verify(verifier)) {
-					return true;
-				}
-			}
-		} catch (Exception e) {
-			e.printStackTrace();
+	public void validateIdTokenByKeys(KeyProperties keyProperties, String idToken) {
+		if (!tokenResolver.verifyPublicKey(keyProperties, idToken)) {
+			throw new JsonParsingException("token.not.valid");
 		}
-
-		return false;
 	}
 }
